adds type component
